Resolved -Wall warning: cast to int for char used as array subscript.

+++ src/reformat/gbtofa/gbtofa.c
@@ -198,7 +198,7 @@ LOOP_START:
         linePt = lineBuf;
         while ((c = *linePt++) != 0)
             {
-            if ((base = ntChars[c]) != 0)
+            if ((base = ntChars[(int)c]) != 0)
                 {
                 if (dnaCount >= dnaAllocated)
                     {
